# Indexing using multiple threads