--- a/Kitodo/src/main/java/org/kitodo/config/enums/ParameterCore.java
+++ b/Kitodo/src/main/java/org/kitodo/config/enums/ParameterCore.java
@@ -621,7 +621,7 @@ public enum ParameterCore implements ParameterInterface {
     ELASTICSEARCH_INDEXLIMIT(new Parameter<>("elasticsearch.indexLimit", 5000)),
     ELASTICSEARCH_BATCH(new Parameter<>("elasticsearch.batch", 500)),
     ELASTICSEARCH_ATTEMPTS(new Parameter<>("elasticsearch.attempts", 10)),
-    ELASTICSEARCH_THREADS(new Parameter<>("elasticsearch.threads", 8)),
+    ELASTICSEARCH_THREADS(new Parameter<>("elasticsearch.threads", 4)),
 
     /*
      * Security properties