--- a/Kitodo/src/main/java/org/kitodo/production/helper/IndexWorker.java
+++ b/Kitodo/src/main/java/org/kitodo/production/helper/IndexWorker.java
@@ -21,15 +21,19 @@ import org.kitodo.config.ConfigCore;
 import org.kitodo.config.enums.ParameterCore;
 import org.kitodo.data.database.exceptions.DAOException;
 import org.kitodo.data.elasticsearch.exceptions.CustomResponseException;
+import org.kitodo.production.enums.ObjectType;
 import org.kitodo.production.services.data.base.SearchService;
 
 public class IndexWorker implements Runnable {
 
-    private int indexedObjects = 0;
-    private int startIndexing;
-    private boolean indexAllObjects = true;
-    private SearchService searchService;
     private static final Logger logger = LogManager.getLogger(IndexWorker.class);
+    
+    private final boolean indexAllObjects;
+    private final ObjectType objectType;
+    private final SearchService searchService;
+    private final IndexWorkerStatus indexWorkerStatus;
+
+    
 
     /**
      * Constructor initializing an IndexWorker object with the given SearchService
@@ -38,103 +42,62 @@ public class IndexWorker implements Runnable {
      * @param searchService
      *            SearchService instance used for indexing
      */
-    public IndexWorker(SearchService searchService) {
+    public IndexWorker(SearchService searchService, ObjectType objectType, IndexWorkerStatus indexWorkerStatus, boolean indexAllObjects) {
         this.searchService = searchService;
-        this.startIndexing = 0;
-    }
-
-    /**
-     * Constructor initializing an IndexWorker object with the given SearchService
-     * and list of objects that will be indexed.
-     *
-     * @param searchService
-     *            SearchService instance used for indexing
-     */
-    public IndexWorker(SearchService searchService, int startIndexing) {
-        this.searchService = searchService;
-        this.startIndexing = startIndexing;
+        this.indexWorkerStatus = indexWorkerStatus;
+        this.indexAllObjects = indexAllObjects;
+        this.objectType = objectType;
     }
 
     @Override
     @SuppressWarnings("unchecked")
     public void run() {
-        this.indexedObjects = 0;
+        int maxAttempts = ConfigCore.getIntParameterOrDefaultValue(ParameterCore.ELASTICSEARCH_ATTEMPTS);
         int batchSize = ConfigCore.getIntParameterOrDefaultValue(ParameterCore.ELASTICSEARCH_BATCH);
-        int indexLimit = ConfigCore.getIntParameterOrDefaultValue(ParameterCore.ELASTICSEARCH_INDEXLIMIT);
-        try {
-            int amountToIndex = getAmountToIndex();
+        int maxBatch = indexWorkerStatus.getMaxBatch();
+        
+        Integer nextBatch = indexWorkerStatus.getAndIncrementNextBatch();
+        while (nextBatch < maxBatch) {
+            // nextBatch is a valid batch that needs to be processed
 
-            if (amountToIndex < batchSize) {
-                if (indexAllObjects) {
-                    indexObjects(searchService.getAll(this.startIndexing, amountToIndex));
-                } else {
-                    indexObjects(searchService.getAllNotIndexed(this.startIndexing, amountToIndex));
-                }
-            } else {
-                if (amountToIndex > indexLimit) {
-                    amountToIndex = indexLimit;
-                }
-                while (this.indexedObjects < amountToIndex) {
-                    indexChunks(batchSize);
+            int attempt = 1;
+            while (attempt < maxAttempts) {
+                try {
+                    int offset = nextBatch * batchSize;
+                    logger.info("index " + objectType.toString() + " with offset " + offset + " and attempt " + attempt);
+
+                    if (indexAllObjects) {
+                        indexObjects(searchService.getAll(offset, batchSize));
+                    } else {
+                        indexObjects(searchService.getAllNotIndexed(offset, batchSize));
+                    }
+
+                    break;
+                } catch (CustomResponseException | DAOException  | HibernateException | IOException e) {
+                    logger.error(e.getMessage(), e);
+                    attempt += 1;
+                    try {
+                        Thread.sleep(2000);
+                    } catch (InterruptedException e2) {
+                        // ignore
+                    }
                 }
             }
-        } catch (CustomResponseException | DAOException  | HibernateException | IOException e) {
-            logger.error(e.getMessage(), e);
+
+            if (attempt >= maxAttempts) {
+                logger.error("stop indexing after maximum amount of attempts");
+                throw new RuntimeException("stop indexing after maximum amount of attempts");
+            }
+
+            // find next batch that can be indexed
+            nextBatch = indexWorkerStatus.getAndIncrementNextBatch();
         }
     }
 
-    private int getAmountToIndex() throws DAOException {
-        if (indexAllObjects) {
-            return searchService.countDatabaseRows().intValue() - this.startIndexing;
-        } else {
-            return searchService.countNotIndexedDatabaseRows().intValue() - this.startIndexing;
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private void indexChunks(int batchSize) throws CustomResponseException, DAOException, IOException {
-        List<Object> objectsToIndex;
-        int indexLimit = ConfigCore.getIntParameterOrDefaultValue(ParameterCore.ELASTICSEARCH_INDEXLIMIT);
-        while (this.indexedObjects < indexLimit) {
-            int offset = this.indexedObjects + this.startIndexing;
-
-            if (indexAllObjects) {
-                objectsToIndex = searchService.getAll(offset, batchSize);
-            } else {
-                objectsToIndex = searchService.getAllNotIndexed(offset, batchSize);
-            }
-            if (objectsToIndex.isEmpty()) {
-                break;
-            }
-
-            indexObjects(objectsToIndex);
-        }
-    }
 
     @SuppressWarnings("unchecked")
     private void indexObjects(List<Object> objectsToIndex) throws CustomResponseException, DAOException, IOException {
         this.searchService.addAllObjectsToIndex(objectsToIndex);
-        this.indexedObjects = this.indexedObjects + objectsToIndex.size();
     }
 
-    /**
-     * Return the number of objects that have already been indexed during the
-     * current indexing process.
-     *
-     * @return int the number of objects indexed during the current indexing run
-     */
-    public int getIndexedObjects() {
-        return indexedObjects + startIndexing;
-    }
-
-    /**
-     * Set value for indexAllObjects. If true, it indexes all objects, if false it
-     * indexes only objects with flag IndexAction.INDEX.
-     *
-     * @param indexAllObjects
-     *            as boolean
-     */
-    public void setIndexAllObjects(boolean indexAllObjects) {
-        this.indexAllObjects = indexAllObjects;
-    }
 }