--- a/Kitodo/src/main/java/org/kitodo/production/helper/IndexWorkerStatus.java
+++ b/Kitodo/src/main/java/org/kitodo/production/helper/IndexWorkerStatus.java
@@ -11,6 +11,7 @@
 
 package org.kitodo.production.helper;
 
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class IndexWorkerStatus {
@@ -28,6 +29,15 @@ public class IndexWorkerStatus {
      */
     private final AtomicInteger nextBatch = new AtomicInteger(0);
 
+    /**
+     * Stores whether there was a failure during indexing.
+     */
+    private final AtomicBoolean failed = new AtomicBoolean(false);
+
+    /**
+     * Whether the indexing process was canceled by a user.
+     */
+    private final AtomicBoolean canceled = new AtomicBoolean(false);
    
     /**
      * Initialize index worker status.
@@ -38,12 +48,53 @@ public class IndexWorkerStatus {
         this.maxBatch = maxBatch;
     }
 
+    /**
+     * Returns the maximum batch number that can be indexed.
+     * 
+     * @return the maximum batch number that can be indexed
+     */
     public int getMaxBatch() {
         return this.maxBatch;
     }
 
+    /**
+     * Access and increase next batch that needs to be indexed in a thread-safe way.
+     * 
+     * @return the next batch that needs to be indexed
+     */
     public int getAndIncrementNextBatch() {
         return this.nextBatch.getAndIncrement();
     }
 
+    /**
+     * Returns whether the indexing process has failed.
+     * 
+     * @return true when the indexing process has failed
+     */
+    public boolean hasFailed() {
+        return failed.get();
+    }
+
+    /**
+     * Marks the index status as failed with some reason.
+     */
+    public void markAsFailed() {
+        this.failed.set(true);
+    }
+
+    /**
+     * Marks the index status as canceld by a user.
+     */
+    public void markAsCanceled() {
+        this.canceled.set(true);
+    }
+
+    /**
+     * Return whether indexing process was canceled by a user.
+     *
+     * @return true if canceled
+     */
+    public boolean isCanceled() {
+        return this.canceled.get();
+    }
 }