… Kitodo/src/main/java/org/kitodo/production/services/index/IndexingService.java
… Kitodo/src/main/resources/kitodo_config.properties
… Kitodo/src/main/webapp/WEB-INF/templates/includes/system/indexing.xhtml
… Kitodo/src/main/webapp/WEB-INF/templates/includes/system/indexingRow.xhtml