This pull request improves the batch indexing performance (triggered via the System/Indexing page) by using multiple threads for indexing. 

A full re-indexing of 81.000 processes and 4.000 tasks took 12 min 12 seconds on my machine (8-core CPU + SSD, default batch size of 250). With these changes, the indexing improves as follows (relative to the best case scenario of the current master branch with a batch size of 2500):

| batch size   | master         | pr threads=1  | pr threads=4  | pr threads=8  |
| --------------- | -------------- | ------------- | ------------- | ------------- |
| 250  | 0.57 (12m:12s) | 1.04 (6m:40s) | 3.39 (2m:03s) | 4.85 (1m:26s) | 
| 500  | 0.76 (9m:12s)  | 0.95 (7m:19s) | 3.50 (1m:59s) | 5.09 (1m:22s) |
| 2500 | **1.00** (6m:57s)  | 1.08 (6m:27s) | 3.26 (2m:08s) | 4.74 (1m:28s) |

There are still a bunch more problems with the indexing, but at least now you will notice them much quicker ;-)

Also, the indexing should be more resilient to intermittent errors. I restarted ElasticSearch during the indexing process and the indexing still finished successfully.

The default number of threads is set to 4 and can be configured via the `elasticsearch.threads` parameter.