@henning-gerhardt 2022-10-05
I tried your changes but you killed the workaround for https://github.com/kitodo/kitodo-production/issues/5179 (refreshing the page to prevent high access rate) and in a short time gigabyte of log entries are created in application proxy in front of the elastic search system.