@thomaslow 2022-10-05
You are probably logging HTTP-requests that go to your ElasticSearch server, right? 

You can limit the number of HTTP-requests used for indexing via the `elasticsearch.batch` parameter. A higher batch size will reduce the number of HTTP-requests required to index all objects, roughly `number_of_objects / batch_size`. 

The HTTP-requests that are used to refresh the current number of already indexed objects (issue #5179) are trigger by the user interface and can be avoided by navigating away from the `System`-page (and not only switching to another tab in the `System`-page), logging out of Kitodo or closing the browser window. The indexing should not stop. As mentioned in #5181, there is currently no button to cancel the indexing process. Alternatively, we could increase the refresh interval to a higher number (e.g. 5 seconds or so?).

Another approach would be to set up log-rotation in your application proxy to avoid huge log files.