@solth 2022-10-10
> I also added a simple cancel link to the UI, since the cancel implementation was very similar to the failure handling.

Does this resolve #5181 ?