@thomaslow 2022-10-10
The cancel link looks like this:

![image](https://user-images.githubusercontent.com/6214043/194871231-51af49a3-7287-47d6-8ba0-4c3d0cdbfb6e.png)

If this is sufficient, then it will resolve #5181.