@henning-gerhardt 2022-10-10
1. Adding new functionality after a review is near done and as new function was not part of any review request or other requirement is not so good in my opinion. I must now find a couple of time in the next days to check the new and existing functions again as no (selenium based) tests are added.
2. If I could say a wish: the new "cancel" button should be appear as the other buttons on this page and should only be visible if an indexing is started.